Cleaning up the abstract syntax tree.
The order of the entries in the abstract syntax tree has been changed
so that it is easier to read and also so that it corresponds exactly
to the order in the formal semantics rules.  This is a first step
towards having complete correspondence between the implementation
and the specification.